Core Düzenlemeler yapıldı

+++ Entites/Concrete/Car.cs
@@ -12,7 +12,7 @@ namespace Entites.Concrete
         public int ColorId { get; set; }
         public string CarName { get; set; }
         public string ModelYear { get; set; }
-        public float DailyPrice{ get; set; }
+        public double DailyPrice{ get; set; }
 
         public string Description { get; set; }
     }
